Moved NotFoundHandler to folder RequestHandler
